build: column resize examples not set up consistently (#19672)
All of the column resize examples were called `*-demo` which is inconsistent with the rest of the examples and will probably break if we decided to expose it to the docs site.

Note that the setup is still a bit inconsistent, because we have different modules for each example that have to be imported separately, but this is due to a limitation in the column resize API where users import a different module depending on whether resizing is enabled by default. I've left a TODO to fix it eventually when we've had the chance to clean up the public API.