fix(sidenav): restore focus if drawer is closed through backdrop click (#19534)
Currently we do not restore focus properly when a drawer is closed
through backdrop click. This is because we only restore focus when
an element inside the drawer is focused before closing. This is not
the case when the backdrop is clicked, as the focus is usually
redirected to the `document.body` element.

Fixes #17749.