fix(sort): sorted state being read out when navigating cells
Reworks the sort header to allow for screen readers to read out its sorted state through `aria-sort`, rather than having to declare an `aria-label` inside it. The drawback of the `aria-label` approach is that the label will also be read out as the user is navigating the table's cells.

Fixes #13012.