test: ignore test failures in popover-edit due to minor subpixel deviations
A small set of popover-edit tests relies on the bounding client rect.
Assertions that the panel stretches over the first and last cell are
correct but require an exact equality. This is not reliable as in some
situations, browsers handle subpixels differently, resulting in
small deviations in the bounding client rect positions. These
differences are only noticeable less significant decimal places, but
the overall assertion matches full pixels. This is sufficient for these
tests so we just round all measurements/comparisons to full pixels.