refactor(cdk/overlay): remove logic which tracks excluded elements from outside clicks (#19887)
This removes the ability to track excluded elements from the overlay and only stops outside
click event notifications if the click event occurred within an overlay element itself. The
overlay should not track elements to be excluded from outside click - this should be up
to the developer. This allows for custom closeout logic to be implemented by the developer leaving
the cdk to be unopinionated.