build: do not leak providers and component styles between tests
Angular's `TestBed` by default only removes a test component from the
document body and destructs individual component fixtures. It never
destroys the test module properly, but always re-creates it.

This means that providers are not necessarily cleaned up (could
leak in memory) and component styles are not cleaned up / deduped.
This results in thousands of style elements in the browsers. Ultimately
causing significantly increased memory consumption and CPU blocking.

This potentially leads to repeated crashes within browsers (as seen in
BrowserStack and Saucelabs in the past). Initial testing of this change
unveiled a reduction from 30min tests to 5min in BrowserStack. This is
a signficant improvement and we should consider moving this change
upstream with: https://github.com/angular/angular/issues/31834.

Benefits are: reduced test time; increased test/browser stability;
isolated tests without leaking styles and services.