fix: add non-null assertions for non-nullable input bindings (#20136)
Due to a bug in the template type checker, binding expressions that include
`undefined` in their type which are bound to directive inputs that don't accept
`undefined` would not be reported as an error, if the directive declares at
least one coercion member using `ngAcceptInputType`. angular/angular#38273
fixes this bug, which uncovers some template issues.

Long-term, it would be best if the directives would be updated to include
`undefined` in an input's type as appropriate. This would require a larger
effort and is likely to introduce some breaking changes to public API.