feat(material-experimental/mdc-autocomplete): implement MDC-based mat-autocomplete (#20247)
* Moves all of the autocomplete logic into base classes so that it can be reused between the standard and MDC components.
* Re-implements `mat-autocomplete` using the logic from the existing one and the styling from MDC.

The MDC-based autocomplete behaves identically to the existing one, with the only minor difference being that MDC one fixes a long-standing issue where we expect a hardcoded height for each of the options. It was easier to fix the bug and add logic to support arbitrary option heights than to add more logic to account for MDC's styles.