build: use new labeling and branching in merge script (#20295)
We introduced a new shared configuration for merge script
labels that follow the new proposal:

https://docs.google.com/document/d/197kVillDwx-RZtSVOBtPb4BBIAw0E9RT3q3v6DZkykU

These label semantics and the new branching are set up for the Angular
Components repository with this commit. The goal is that labeling and
merging is consistent between all Angular projects and that clear rules
are defined for branching/labeling. This was previously not the case