feat(cdk-experimental/menu): add smart aim functionality to the menu (#20442)
* feat(cdk-experimental/menu): add smart aim functionality to the menu

Implement Smart Aim which ensures that when a user is mousing through menus, the target
submenu the user is mousing into isn't closed prematurely. Currently, a user must perform precise
and deliberate mouse moves in order to avoid sibling trigger items which close the target submenu
the user is intending to go into.

* feat(cdk-experimental/menu): make TargetMenuAim optional via a dedicated provider

Enable TargetMenuAim to be optionally injected when the cdkTargetMenuAim directive is placed next
to cdkMenu or cdkMenuBar.

* refactor(cdk-experimental): change smart aim algo to use direct slope ignoring layout

Change the Smart Aim algorithm to track the last n points and determine a consensus between them
when deciding if a user is moving towards the submenu.

* refactor(cdk-experimental/menu): change Smart Aim logic to check for bounding points

Update the Smart Aim to check if any point on the users mouse trajectory line falls within the
bounds of the submenu rather than calculating at left, middle, right