mdc-list: support for strong focus indicators and fix incorrect cursor for non-interactive items (#20476)
* feat(material-experimental/mdc-list): add support for strong focus indicators

Adds support for strong focus indicators to the MDC-based list.
Also simplifies SCSS code for the hover/focus/selected styles.

* fix(material-experimental/mdc-list): do not show pointer cursor for non-interactive list items

The MDC list currently always shows the pointer cursor for all kinds
of list items, regardless of them being interactive or not (through
instantiation of the foundation).

This seems to be unexpected as the pointer cursor incorrectly suggests
interactivity of a list item to users. We manually reset the cursor
for non-interactive list items and report an issue upstream:

https://github.com/material-components/material-components-web/issues/6443

Note: MDC list does no longer seem to support the `non-interactive`
class we previously used, so we replace it with our own.