fix(material-experimental/mdc-list): checkbox and icon alignment incorrect
The current logic for checkbox and icon alignment is not correct. Icons
are not displayed correctly due to mismatching conditionals for the template
bindings. This commit fixes these bindings and simplifies the logic slightly.

Additionally, the way icons and avatars are rendered within options has
been changed. They are not nested inside any of the graphic/meta
containers but instead are projected as direct children of the
list-option. This is necessary for MDC list items so that icons/avatars
that already have the proper meta/graphic class align properly
horizontally and vertically.

Note: Initially, the checkbox and icons were conditionally rendered as part
of a single template outlet. This broke with View Engine so we looked
into manually managing the embedded views. This worked as expected but
resulted in inconvenient code that we'd need to maintain and ship to
users. After more experiments and comparisons with Ivy, it looks like we
can overcome this issue by now just leveraging two template outlets for
either icons and the checkbox.