refactor: remove light-weight token content-child query cast (#20433)
Initially when we added light-weight tokens to our components,
the Angular framework was not properly typed to support
injection tokens in the view and content queries.

We fixed this upstream with https://github.com/angular/angular/pull/37506
and can now remove the workaround/cast since this change is available
due to our recent Angular update to a `next` version.